+++ packages/core-js-compat/src/data.js
@@ -1247,6 +1247,8 @@ const data = {
   },
   'esnext.iterator.find': {
   },
+  'esnext.iterator.flat-map': {
+  },
   'esnext.iterator.for-each': {
   },
   'esnext.iterator.from': {